# .well-known/openapi.yaml
openapi: 3.0.2
info:
  title: OpenAI plugin for a simple todo app
  description: Todo app for managing your tasks on ChatGPT
  version: 1.0.0
  servers:
    - url: https://minsa110-laughing-eureka-979q457rx6q3p954-8000.preview.app.github.dev
paths:
  /todos:
    post:
      summary: Create a new TODO item
      description: Accepts a string and adds as new TODO item
      operationId: create_todo
      parameters:
        - in: query
          name: todo
          schema:
            type: string
          required: true
          description: The description of the TODO item
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoItem"
    get:
      summary: Get a list of all TODO items
      operationId: list_todos
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TodoList"
  /todos/{todo_id}:
    get:
      summary: Get a TODO item by ID
      operationId: get_todo
      parameters:
        - name: todo_id
          in: path
          required: true
          description: ID of the TODO item to retrieve
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TodoItem"
        "404":
          description: Todo not found
    delete:
      summary: Delete a TODO item by ID
      operationId: delete_todo
      parameters:
        - name: todo_id
          in: path
          required: true
          description: ID of the TODO item to delete
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Todo deleted
        "404":
          description: Todo not found
components:
  schemas:
    TodoItem:
      type: object
      properties:
        todo:
          type: string
        todo_id:
          type: integer
          format: int32
          readOnly: true
      required:
        - todo